refactor statement
